refactor: remove useless files in example apps and unify source
