--- apps/examples/index.js
@@ -1,3 +1,0 @@
-import App from './App';
-
-export default App;
