+++ apps/index.tsx
@@ -1,0 +1,3 @@
+import App from './examples';
+
+export default App;
