--- fabric-example/App.tsx
@@ -1,3 +1,0 @@
-import App from '../apps/examples';
-
-export default App;
