--- tests-example/App.js
@@ -1,3 +1,0 @@
-import App from '../apps/test-examples';
-
-export default App;
