+++ tests-example/index.js
@@ -2,8 +2,8 @@
  * @format
  */
 
-import {AppRegistry} from 'react-native';
-import App from './App';
-import {name as appName} from './app.json';
+import { AppRegistry } from 'react-native';
+import App from '../apps/test-examples';
+import { name as appName } from './app.json';
 
 AppRegistry.registerComponent(appName, () => App);
